HT_24/requirements.txt
Django==5.0.1
fake-useragent==1.4.0
requests==2.31.0
djangorestframework==3.14.0
django-filter==23.5
celery==5.3.6
redis==5.0.1
pytest==8.0.0
factory_boy==3.3.0